[FIX] Add "montant" field
